HuggingFaceApi

Documentation for HuggingFaceApi.

HuggingFaceApi.HuggingFaceURL

HuggingFaceApi.cached_download

HuggingFaceApi.hf_hub_download

HuggingFaceApi.huggingface_emoji

HuggingFaceApi.login

HuggingFaceApi.logout

HuggingFaceApi.remove_cache

HuggingFaceApi.HuggingFaceURL — Type

HuggingFaceURL(repo_id, [subfolder], filename,
               repo_type::REPO_TYPE = nothing,
               revision::AbstractString = "main")

Construct the real url with the inputs.

HuggingFaceApi.cached_download — Method

cached_download(
    hgfurl :: HuggingFaceURL;
    local_files_only :: Bool = false,
    auth_token :: Union{AbstractString, Nothing} = nothing,
)

Find the local cache of given url or do downloading. If local_files_only is set, it will try to find the file from cache, and error out when not found. For downloading from private repo, auth_token need to be set, or do HuggingFaceApi.login() beforehand.

See also: HuggingFaceURL, login

HuggingFaceApi.hf_hub_download — Method

hf_hub_download(
    repo_id :: AbstractString,
    filename :: AbstractString;
    repo_type :: REPO_TYPE = nothing,
    revision :: AbstractString = "main",
    auth_token :: Union{AbstractString, Nothing} = get_token(),
    local_files_only :: Bool  = false,
)

Construct HuggingFaceURL and do cached_download.

HuggingFaceApi.huggingface_emoji — Method

add 🤗 autocomplete in REPL

HuggingFaceApi.login — Method

login(user::AbstractString = readline())

Login to huggingface hub and get/save the account token for operations that need authentication. Re-run this function with username (i.e. login(username)) if token expired.

HuggingFaceApi.logout — Function

logout()

Login from huggingface hub and remove all authentication cache.

HuggingFaceApi.remove_cache — Method

remove_cache(hgfurl::HuggingFaceURL; now=false)

Remove files link to the given url. If now is set to true, cache file will be deleted immediately, otherwise waiting OhMyArtifacts to do the garbage collection.